refactor: use storage term instead of persistence
