+++ apps/chrome-extension/src-extension/background.ts
@@ -1,6 +1,6 @@
 /* global chrome  */
 import { broadcastMessage } from '@/chrome/utils'
-import { STORAGE_KEY_SETTINGS } from '@/persistence'
+import { STORAGE_KEY_SETTINGS } from '@/storage'
 
 chrome.runtime.onInstalled.addListener(async () => {
   console.log('readapt installed')
