+++ apps/chrome-extension/src-extension/readapt.ts
@@ -1,7 +1,7 @@
 /* global chrome */
 import { maskState, rulerState, addMask, addRuler, updateRulerSettings, removeRuler, removeMask, updateMaskSettings } from './reading-tools'
 import { loadVisualEngine } from './load-visual-engine'
-import { STORAGE_KEY_SETTINGS } from '@/persistence'
+import { STORAGE_KEY_SETTINGS } from '@/storage'
 
 const mousePos = {
   x: 0,
