--- apps/chrome-extension/src/persistence/index.ts
@@ -1,3 +1,0 @@
-export * from './constants'
-export { Persistence } from './persistence'
-export { PersistenceMock } from './persistenceMock'
