+++ apps/chrome-extension/src/storage/index.ts
@@ -1,0 +1,3 @@
+export * from './constants'
+export { Storage } from './storage'
+export { StorageMock } from './storageMock'
