+++ apps/chrome-extension/src/storage/storage.ts
@@ -1,4 +1,4 @@
-export interface Persistence {
+export interface Storage {
   getItem<Item>(key: string): Promise<Item | undefined>
   setItem<Item>(key: string, item: Item): Promise<void>
 }
